+++ users/models/__init__.py
@@ -1,0 +1,22 @@
+# Re-export all models for backwards compatibility with `from users.models import ...`
+
+from .account import CustomUser, ActiveUserManager
+from .profile import Profile, LanguageProficiency, ActiveProfileManager
+from .social import SavedItem, LetsConnectLog
+from .badges import Badge, UserBadge
+from .reference import (
+    Territory,
+    Language,
+    WikimediaProject,
+    Avatar,
+    AuthExtraInfo,
+    DataHash,
+)
+
+__all__ = [
+    'CustomUser', 'ActiveUserManager',
+    'Profile', 'LanguageProficiency', 'ActiveProfileManager',
+    'SavedItem', 'LetsConnectLog',
+    'Badge', 'UserBadge',
+    'Territory', 'Language', 'WikimediaProject', 'Avatar', 'AuthExtraInfo', 'DataHash',
+]
